Move `FILE` to `stdio`
